reseduial/requirements-minimal.txt
# Minimal HealthPredict Requirements
# Install with: pip install -r requirements-minimal.txt

# Core Web Framework
fastapi==0.116.1
uvicorn==0.30.0

# Streamlit Framework (for prediction apps)
streamlit==1.41.1

# Machine Learning & Data Processing
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2

# API Integration
openai==1.59.8

# Environment Management
python-dotenv==1.0.1

# Optional but Recommended for Production
gunicorn==23.0.0
